pyproject.toml
[tool.poetry]
name = "atomichabits"
version = "0.1.0"
description = ""
authors = ["boshy <kaidohmary@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.7"
celery = "^5.3.4"
redis = "^5.0.1"
pytelegrambotapi = "^4.14.0"
django-celery-beat = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
